Move the parity helper method to private ElementParity trait
